Add crud for PriceHistory, view last price

+++ src/main/java/ru/job4j/cars/service/SimplePostService.java
@@ -10,16 +10,14 @@ import ru.job4j.cars.mapstruct.PostMapper;
 import ru.job4j.cars.mapstruct.UserMapper;
 import ru.job4j.cars.model.Car;
 import ru.job4j.cars.model.Post;
+import ru.job4j.cars.model.PriceHistory;
 import ru.job4j.cars.repository.CarRepository;
 import ru.job4j.cars.repository.PostRepository;
 import ru.job4j.cars.repository.UserRepository;
 import ru.job4j.cars.util.Utils;
 
 import javax.persistence.criteria.CriteriaBuilder;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,9 +26,9 @@ public class SimplePostService implements PostService {
 
     private final PostRepository postRepository;
 
-    private final CarRepository carRepository;
-
     private final CarService carService;
+
+    private final PriceHistoryService priceHistoryService;
 
     private final PostMapper postMapper;
 
@@ -40,39 +38,56 @@ public class SimplePostService implements PostService {
     public PostDto create(PostDto postDto, UserDto userDto, CarDto carDto) {
         Post post = postMapper.getEntityFromDto(postDto);
         post = postRepository.create(post);
-        return postMapper.getModelFromEntity(post, carDto, 10000);
+        var priceHistory = priceHistoryService.create(post, postDto.getPrice());
+        return postMapper.getModelFromEntity(post, carDto, priceHistory);
     }
 
     @Override
     public Optional<PostDto> findById(int id, UserDto userDto) {
-        Optional<Post> optionalPost = postRepository.findById(id);
+        //TODO Упростить или каскадом подтягивать
         Optional<PostDto> optionalPostDto = Optional.empty();
+        Optional<Post> optionalPost = postRepository.findById(id);
         if (optionalPost.isEmpty()) {
             return optionalPostDto;
         }
         var post = optionalPost.get();
         post = Utils.correctTimeZone(post, userDto.getTimezone());
+
         Optional<CarDto> optionalCarDto = carService.findById(post.getCar().getId());
-        optionalPostDto = Optional.of(postMapper.getModelFromEntity(post, optionalCarDto.get(), 10000));
+        var carDto = optionalCarDto.orElseGet(CarDto::new);
+
+        Optional<PriceHistory> optionalPriceHistory = priceHistoryService.findLastByPostId(id);
+        var priceHistory = optionalPriceHistory.orElseGet(PriceHistory::new);
+
+        optionalPostDto = Optional.of(postMapper.getModelFromEntity(post, carDto, priceHistory));
         return optionalPostDto;
     }
 
     @Override
     public List<PostDto> findAll(UserDto userDto) {
         List<Post> posts = postRepository.findAllOrderById();
+
         List<CarDto> carDtos= carService.findAll();
         Map<Integer, CarDto> mapAllCarDtos = carDtos.stream().
                 collect(Collectors.toMap(CarDto::getId, carDto -> carDto));
+
+        List<PriceHistory> priceHistories = priceHistoryService.findAllLastPrice();
+        Map<Integer, PriceHistory> mapAllPH = new HashMap<>();
+        for (PriceHistory ph : priceHistories) {
+            mapAllPH.put(ph.getPost().getId(), ph);
+        }
+
         List<PostDto> postDtos = new ArrayList<>();
         for (Post post : posts) {
             post = Utils.correctTimeZone(post, userDto.getTimezone());
-            postDtos.add(postMapper.getModelFromEntity(post, mapAllCarDtos.get(post.getCar().getId()), 10000));
+            postDtos.add(postMapper.getModelFromEntity(post, mapAllCarDtos.get(post.getCar().getId()), mapAllPH.get(post.getId())));
         }
         return postDtos;
     }
 
     @Override
     public boolean update(PostDto postDto, UserDto userDto) {
+        priceHistoryService.update(postDto, postDto.getPrice());
         return postRepository.update(postMapper.getEntityFromDto(postDto), userMapper.getEntityFromDto(userDto));
     }
 
@@ -83,6 +98,9 @@ public class SimplePostService implements PostService {
 //
     @Override
     public boolean deleteById(int id, UserDto userDto) {
+        if (!priceHistoryService.deleteAllByPostId(id)) {
+            return false;
+        }
         return postRepository.delete(id, userMapper.getEntityFromDto(userDto));
     }
 }
